04_Serverless_RAG_Assistant/requirements.txt
# Core dependencies
python>=3.11
boto3>=1.28.0
aws-cdk-lib>=2.100.0
constructs>=10.0.0

# LangChain (optional, for full RAG implementation)
# langchain>=0.1.0
# langchain-openai>=0.0.5
# langchain-community>=0.0.20

# Document processing
PyPDF2>=3.0.0
python-docx>=1.1.0

# Vector store (optional)
# chromadb>=0.4.0
# faiss-cpu>=1.7.4

# API
fastapi>=0.104.0
uvicorn>=0.24.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

